+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Use latest Alpine Linux 3.x for Alpine-based docker containers instead of always 3.14.
 * Updated CentOS 7 dockerfiles to vault repo due to EOL.
 * Use number of parallel jobs from makefile environment variable for builds of external software in `prepare-external.sh`.
+* Reseed golden prime based random number generator after 256KB at the latest with a random number generator from a stronger generator.
+* Rotate through multiple golden primes for fast random number generator.
 
 ### Fixes
 * Fixed retry mechanism for S3 object download in case server is temporarily unavailable.